requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
pypdf2>=3.0.0
langchain>=0.0.300
langchain-community>=0.0.5
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rank-bm25>=0.2.0
redis>=5.0.0
python-multipart>=0.0.5
aiofiles>=23.0.0
